requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
drf-yasg>=1.21
django-cors-headers>=4.0
psycopg2-binary>=2.9